backend/requirements.txt
Django==1.9.8
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.6.0
Unipath==1.0
python-decouple==3
Pillow==3.3.0
Markdown==2.6.6
bleach==1.4.3
psycopg2==2.6.1
whitenoise==3.2.2
django-filter==0.11.0
djangorestframework==3.3.1
virtualenv==15.0.2
django-allauth
requests 
requests-oauthlib
pyfcm==1.0.5
django-rest-auth
